backend/hackathon/requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
Django==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.13.1
drf-spectacular==0.23.1
geographiclib==1.52
geopy==2.2.0
h11==0.13.0
idna==3.3
inflection==0.5.1
jsonschema==4.9.1
newsapi-python==0.2.6
numpy==1.23.2
outcome==1.2.0
pandas==1.4.3
Pillow==9.2.0
pycparser==2.21
pyOpenSSL==22.0.0
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
reportlab==3.6.12
requests==2.28.1
scipy==1.9.0
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.11
wsproto==1.1.0
